+++ src/client/js/index.js
@@ -2,6 +2,8 @@
  * pwix:tabular/src/client/js/index.js
  */
 
+import { default as alTabular } from 'meteor/aldeed:tabular';
+
 // import here templates addressed by the  Tabular class
 import '../components/dt_checkbox/dt_checkbox.js';
 import '../components/dt_buttons/dt_buttons.js';
@@ -10,6 +12,7 @@ import '../components/dt_buttons/dt_buttons.js';
 import '../../common/js/index.js';
 
 // mutualize the datatable imports
+import DataTables from 'datatables.net-bs5';
 import 'datatables.net-buttons-bs5';
 import 'datatables.net-buttons/js/buttons.colVis.mjs';
 import 'datatables.net-buttons/js/buttons.print.mjs';
@@ -28,3 +31,5 @@ import 'datatables.net-rowgroup-bs5/css/rowGroup.bootstrap5.min.css';
 import 'datatables.net-scroller-bs5/css/scroller.bootstrap5.min.css';
 //
 import '../stylesheets/tabular.less';
+//
+alTabular.init({ DataTables });